Move files to src subdirs
